Update default variant to aws-k8s-1.21
